refactor: rename TopicsTab.tsx
